Store: make index cache shared cache friendly (#1825)
* Renamed store cache.go to inmemory.go


* Refactored IndexCache to multiple fetch postings and series within a single call


* Added missing comments to exported functions from InMemoryIndexCache
